Rationalize redirect rendering in AuthorizationEndpoint
There are 3 methods in AuthorizationEndpoint that need to build a
redirect URI, and up to now there were 3 different methods of
doing it. This change unifies them into a single convenience method.

It also allows the incoming redirect URI to be either encoded or
unencoded (fixes gh-349). An unencoded URI can easily be provided
as part of a ClientDetails registration, and an encoded one would
come in as a request parameter. (Actually request parameters can
be unencoded too, but most clients will encode to avoid
ambiguity.)